requirements-dev.txt
-r requirements.txt
Django==4.2.4
requests
hiredis
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-ckeditor
PyYAML
uritemplate
Pygments
django-guardian
psycopg2-binary
pillow
whoosh
django-haystack
jieba
django-simpleui
django-import-export
django-debug-toolbar
oss2
wechatpayv3
django-compressor
sentry-sdk[django]
django-celery-results
django-celery-beat